umu: move argument parsing to __main__.py
